pmdabpf: fix total metric indom variable type signedness
Minor cleanup after reviewing previous fix in this code.

+++ src/pmdas/bpf/bpf.c
@@ -291,8 +291,8 @@ void
 bpf_register_module_metrics()
 {
     // identify how much space we need and set up metric table area
-    int total_metrics = 0;
-    int total_indoms = 0;
+    size_t total_metrics = 0;
+    size_t total_indoms = 0;
     int cache_op_status;
     module* bpf_module;
     char indom[64];
